requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
python-decouple==3.8
django-cors-headers==4.3.1
django-filter==25.1

# Celery and Task Queue
celery==5.3.4
redis==5.0.1
flower==2.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.0

# System Monitoring
psutil==5.9.6

# Testing
factory-boy==3.3.0
pytest-django==4.5.2
pytest-cov==4.1.0

# Production Dependencies
whitenoise==6.6.0
django-redis==5.4.0

# Optional PostgreSQL support (requires Visual C++ Build Tools on Windows)
# psycopg2-binary==2.9.7

# Optional Monitoring (Production)
# sentry-sdk==1.38.0